+++ package.json
@@ -54,11 +54,6 @@
           "command": "lengalab.setStructuredView",
           "when": "activeCustomEditorId == lengalab.c",
           "group": "navigation"
-        },
-        {
-          "command": "lengalab.setSpikeView",
-          "when": "activeCustomEditorId == lengalab.c",
-          "group": "navigation"
         }
       ]
     }
