+++ rpc/generate.sh
@@ -7,7 +7,7 @@ set -euo pipefail
 # ================================
 GITHUB_USER="lengalab"                      # GitHub username or org
 GITHUB_REPO="lenga"                         # Repository name
-BRANCH_OR_COMMIT="feat/rust-proto-translation"   # Can be branch or pinned commit hash
+BRANCH_OR_COMMIT="feat/save-service"   # Can be branch or pinned commit hash
 REMOTE_DIRECTORY="lenga-server/proto"       # Directory in the repo you want
 PROTOS_DIR="rpc/protos"                     # Relative to git repository root
 GENERATED_DIR="rpc/generated"               # Relative to git repository root
