--- webview-ui/src/spikeView/index.css
@@ -1,78 +1,0 @@
-*,
-::before,
-::after {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-html,
-body,
-#root {
-  height: 100%;
-  width: 100%;
-}
-
-.inputNode {
-  display: flex;
-  gap: 0.5rem;
-}
-
-.inputNode input {
-  width: 5rem;
-}
-
-.inputNode button {
-  background-color: transparent;
-  border: solid 1px black;
-  padding: 0 0.5rem;
-  border-radius: 0.1rem;
-}
-
-.textNode {
-  width: 8rem;
-  height: 3rem;
-  text-align: center;
-  border-radius: 0.5rem;
-  background-color: white;
-  padding: 0.5rem;
-  border: solid 1px black;
-  overflow-x: auto;
-  overflow-y: hidden;
-}
-
-.function-node {
-  border: 2px solid #000000;
-  border-radius: 8px;
-  padding: 8px;
-  background-color: #f9f9f9;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
-}
-
-.editable-textarea {
-  font-size: 14px;
-  font-family: inherit;
-  color: #333;
-  line-height: 1.4;
-  padding: 0;
-  border: none;
-  outline: none;
-  resize: none;
-  overflow: hidden;
-  background: transparent;
-  white-space: pre;
-  min-width: 120px;
-}
-
-.textarea-mirror {
-  position: absolute;
-  top: -9999px;
-  left: 0;
-  white-space: pre-wrap;
-  word-wrap: break-word;
-  visibility: hidden;
-  font-size: 14px;
-  font-family: inherit;
-  line-height: 1.4;
-  padding: 0;
-}
